+++ src/main/java/com/global/api/paymentMethods/CreditCardData.java
@@ -93,12 +93,24 @@ public class CreditCardData extends Credit implements ICardData {
         return !StringUtils.isNullOrEmpty(this.getToken()) && this.getMobileType() != null;
     }
 
+    /**
+     * @deprecated verifyEnrolled is deprecated. Please use CheckEnrollment from Secure3dService
+     */
+    @Deprecated
     public boolean verifyEnrolled(BigDecimal amount, String currency) throws ApiException {
         return verifyEnrolled(amount, currency, null, "default");
     }
+    /**
+     * @deprecated verifyEnrolled is deprecated. Please use CheckEnrollment from Secure3dService
+     */
+    @Deprecated
     public boolean verifyEnrolled(BigDecimal amount, String currency, String orderId) throws ApiException {
         return verifyEnrolled(amount, currency, orderId, "default");
     }
+    /**
+     * @deprecated verifyEnrolled is deprecated. Please use CheckEnrollment from Secure3dService
+     */
+    @Deprecated
     public boolean verifyEnrolled(BigDecimal amount, String currency, String orderId, String configName) throws ApiException {
         Transaction response;
         try{
@@ -127,9 +139,15 @@ public class CreditCardData extends Credit implements ICardData {
         return false;
     }
 
+    /**
+     * @deprecated verifySignature is deprecated. Please use CheckEnrollment from Secure3dService
+     */
     public boolean verifySignature(String authorizationResponse, BigDecimal amount, String currency, String orderId) throws ApiException {
         return verifySignature(authorizationResponse, amount, currency, orderId, "default");
     }
+    /**
+     * @deprecated verifySignature is deprecated. Please use CheckEnrollment from Secure3dService
+     */
     public boolean verifySignature(String authorizationResponse, BigDecimal amount, String currency, String orderId, String configName) throws ApiException {
         // ensure we have an object
         if(this.threeDSecure == null)
@@ -141,12 +159,21 @@ public class CreditCardData extends Credit implements ICardData {
 
         return verifySignature(authorizationResponse, null, configName);
     }
+    /**
+     * @deprecated verifySignature is deprecated. Please use CheckEnrollment from Secure3dService
+     */
     public boolean verifySignature(String authorizationResponse) throws ApiException {
         return verifySignature(authorizationResponse, null, "default");
     }
+    /**
+     * @deprecated verifySignature is deprecated. Please use CheckEnrollment from Secure3dService
+     */
     public boolean verifySignature(String authorizationResponse, MerchantDataCollection merchantData) throws ApiException {
         return verifySignature(authorizationResponse, merchantData, "default");
     }
+    /**
+     * @deprecated verifySignature is deprecated. Please use CheckEnrollment from Secure3dService
+     */
     public boolean verifySignature(String authorizationResponse, MerchantDataCollection merchantData, String configName) throws ApiException {
         // ensure we have an object
         if(threeDSecure == null) {